youtubefilter/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
psycopg2-binary==2.9.1
gunicorn==20.1.0
argon2-cffi>=21.1.0
djangorestframework-simplejwt==5.0.0
django-filter==21.1
requests==2.26.0